+++ lib/hexapdf/type/document_security_store.rb
@@ -34,6 +34,7 @@
 # commercial licenses are available at <https://gettalong.at/hexapdf/>.
 #++
 
+require 'openssl'
 require 'hexapdf/dictionary'
 
 module HexaPDF
@@ -69,6 +70,63 @@ module HexaPDF
       define_field :Certs, type: PDFArray, version: '2.0'
       define_field :OCSPs, type: PDFArray, version: '2.0'
       define_field :CRLs,  type: PDFArray, version: '2.0'
+
+      # Adds validation data for a single signature to the /VRI dictionary and to the various
+      # arrays, and returns the new VRI entry.
+      #
+      # +signature+::
+      #     The signature for which the validation data should be added.
+      #
+      # +certs+::
+      #     Array of DER-encoded certificates.
+      #
+      # +ocsps+::
+      #     Array of DER-encoded OCSP responses.
+      #
+      # +crls+::
+      #     Array of DER-encoded CRLs.
+      def add_vri(signature, certs: [], ocsps: [], crls: [])
+        key = OpenSSL::Digest::SHA1.hexdigest(signature.contents).upcase.to_sym
+        vri = {Type: :VRI}
+        vri[:Cert] = certs.map {|der| add_cert(der) } unless certs.empty?
+        vri[:OCSP] = ocsps.map {|der| add_ocsp(der) } unless ocsps.empty?
+        vri[:CRL]  = crls.map  {|der| add_crl(der)  } unless crls.empty?
+        (self[:VRI] ||= {})[key] = document.add(vri)
+      end
+
+      # Adds the DER-encoded certificate to the /Certs array if not already present and returns
+      # the stream object containing it.
+      def add_cert(cert_der)
+        add_data_as_stream_to_field(cert_der, :Certs)
+      end
+
+      # Adds the DER-encoded OCSP response to the /OCSPs array if not already present and returns
+      # the stream object containing it.
+      def add_ocsp(ocsp_der)
+        add_data_as_stream_to_field(ocsp_der, :OCSPs)
+      end
+
+      # Adds the DER-encoded CRL to the /CRLs array if not already present and returns the stream
+      # object containing it.
+      def add_crl(crl_der)
+        add_data_as_stream_to_field(crl_der, :CRLs)
+      end
+
+      private
+
+      # Adds the given string +data+ to the +field+ array as stream object and returns the resulting
+      # stream.
+      #
+      # If the field already contains a stream with the same data, this existing stream is returned.
+      def add_data_as_stream_to_field(data, field)
+        self[field] ||= []
+        existing_stream = self[field].find {|stream| stream.stream == data }
+        return existing_stream if existing_stream
+
+        stream = document.add({Filter: :FlateDecode}, stream: data)
+        self[field] << stream
+        stream
+      end
 
     end